requirements.txt
adafruit-io==2.6.0
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
mysql-connector-python==8.0.28
paho-mqtt==1.6.1
protobuf==3.19.4
pyFirmata==1.1.0
pyserial==3.5
requests==2.27.1
urllib3==1.26.8
